I fixed problem with misaligned compression. 24 mod 8 is zero, so the Y-channel was fine, but the color channels are downsampled, so the offset for collecting samples from low-pass images needs to be a factor of 16, not 8. Code now passes alignment stupidity test: the compressed image is indeed the sum of the low-pass image and the high-pass image.

+++ generate_low_freq_imgs.py
@@ -31,7 +31,7 @@ fftSz1 = (480,320)
 fftSz2 = (320,480)
 smooth_jpeg1 = jrf.Smooth_JPEG(rho,alpha,noi,qY,qUV,lmbda,fftSz1,dtype=dtype)
 smooth_jpeg2 = jrf.Smooth_JPEG(rho,alpha,noi,qY,qUV,lmbda,fftSz2,dtype=dtype)
-Yoffset = tf.one_hot([[[0]]],64,tf.cast(32.,self.dtype),tf.cast(0.,self.dtype))
+Yoffset = tf.one_hot([[[0]]],64,tf.cast(32.,dtype = dtype),tf.cast(0.,dtype= dtype))
 # Loop through images
 dataPath = 'data/original/simpleTest/'
 filelist = os.listdir(dataPath)
